+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod file_cache;
 pub mod filesystem;
 pub mod render;
 pub mod templates;
+pub mod template_helpers;
 pub mod utils;
 pub mod webserver;
